Add custom build backend to support build args
This implements a custom build backend, inspired by [the solution used
by Pillow](https://github.com/python-pillow/Pillow/pull/7171).

With this change, argument flags now need to be specified as key-value
pairs instead of key-only as `--config-settings`. Every argument can
either be left unspecified or have a value that is either "enabled" or
"disabled".

For example:

```
--config-settings --global-option=--cpp_ext
--global-option --cpp_ext
```

becomes

```
--config-settings cpp_ext=enabled
```

The setuptools version requirement of 40.8.0 was also adopted from
Pillow. It is the version that introduced the `build_meta:__legacy__`
backend.